requirements.txt
bottleneck==1.5.0
einops==0.8.1
fire==0.7.0
gin_config==0.5.0
matplotlib==3.7.2
numpy==1.24.1
pandas==2.3.1
scikit_learn==1.3.0
scipy==1.16.1
seaborn==0.13.2
torch==2.0.0+cu118
tqdm==4.66.1
tslearn==0.6.3
